Adapter le frontend pour le backend Node.js/MySQL - dateEntretien et ID number

+++ src/app/core/list-jardin/list-jardin.component.ts
@@ -30,7 +30,8 @@ export class ListJardinComponent {
     });
   }
 
-  deleteJardin(id: string){
+  deleteJardin(id: number | undefined){
+    if (!id) return;
     this.jardinService.deleteJardin(id).subscribe(()=>{
       this.jardins = this.jardins.filter(j => j.id != id);
     });
